fix(tests): complete int semantics in 5 financial test suites
FASE I (Parte 1): 8 fixes simples en tests.

CAMBIOS:

1. CashierEntitiesTest (2 fixes):
   - balanceImpact() retorna int (FASE E)
   - expects: -30000.0 → -30000, 20000.0 → 20000

2. CashSessionBugfixTest (2 fixes):
   - calculateExpectedAmountForClose() retorna int (FASE E)
   - expects: 85000.0 → 85000, 43000.0 → 43000

3. FinancialRulesTest (1 fix):
   - casteo (float) → (int) en suma de payments

4. LedgerTest (1 fix):
   - getAccountBalance() retorna int (FASE F)
   - expect: 7000.0 → 7000

5. PaymentLedgerIntegrationTest (2 fixes):
   - balances de cuentas retornan int (FASE F)
   - expects: 5000.0 → 5000, 7900.0 → 7900, 4000.0 → 4000

TOTAL: 8 expects corregidos en 5 archivos

Siguiente: Corregir OrderItemController.php (genera DECIMAL en tax_amount)

+++ tests/Feature/FinancialRulesTest.php
@@ -364,7 +364,7 @@ test('CASO 7: Múltiples pagos ($5,000 + $3,000 + $2,000 = $10,000)', function (
 
     // Verificar suma
     $totalPaid = Payment::where('order_id', $order->id)->sum('amount');
-    expect((float) $totalPaid)->toBe(10000);
+    expect((int) $totalPaid)->toBe(10000);  // ADR-018: int
 });
 
 // ═══════════════════════════════════════════════════
